SCRUM-118 refactor: implement ProfileInfo, edit types
