+++ src/app/profile/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Profile } from '@/entities/profile'
+import { Profile } from '@/entities/profile/ui'
 
 export default function ProfilePage() {
   return <Profile />
