+++ src/app/profile/page.tsx
@@ -1,9 +1,9 @@
 'use client'
-import { useAuth } from '@/features/posts/utils/useAuth'
+import { useMeQuery } from '@/features/auth'
 import { redirect } from 'next/navigation'
 
 export default function ProfileRedirect() {
-  const { user } = useAuth()
+  const { data: user } = useMeQuery()
 
   if (!user) {
     redirect('/')
