--- src/entities/profile/index.ts
@@ -1,3 +1,0 @@
-export * from './ui'
-export * from './api'
-export * from './hooks'
