+++ src/entities/profile/ui/Profile/ProfileBio/ProfileBio.tsx
@@ -1,11 +1,9 @@
 import { Typography } from '@/shared/ui'
 
 type Props = {
-    message: string | null
+  message: string | null
 }
 
 export const ProfileBio = ({ message }: Props) => (
-    <Typography variant="regular_16" >
-        {message || 'No information has been added yet.'}
-    </Typography>
-);
+  <Typography variant="regular_16">{message || 'No information has been added yet.'}</Typography>
+)
