+++ src/entities/profile/ui/Profile/ProfileInfo/ProfileInfo.module.scss
@@ -1,0 +1,14 @@
+.info {
+  display: flex;
+  gap: 38px;
+}
+
+.details {
+  width: 100%;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 20px;
+}
