+++ src/entities/profile/ui/Profile/ProfileInfo/index.ts
@@ -1,0 +1,1 @@
+export * from './ProfileInfo'
