+++ src/entities/profile/ui/Profile/ProfilePosts/ProfilePosts.module.scss
@@ -1,3 +1,8 @@
+.wrapper {
+  display: flex;
+  text-align: center;
+}
+
 .posts {
   width: 100%;
   display: grid;
@@ -6,5 +11,6 @@
 }
 
 .message {
+  text-align: center;
   width: 100%;
 }
