+++ src/entities/profile/ui/Profile/index.ts
@@ -3,4 +3,4 @@ export * from './ProfileActions'
 export * from './ProfileBio'
 export * from './ProfilePosts'
 export * from './ProfileStats'
-export * from './InfiniteScrollTrigger'
+export * from './ProfileInfo'
