+++ src/shared/composites/InfiniteScrollTrigger/index.ts
