+++ src/shared/composites/index.ts
@@ -9,3 +9,4 @@ export * from './InfoPage'
 export * from './Avatar'
 export * from './Carousel'
 export * from './ActionsMenu'
+export * from './InfiniteScrollTrigger'
