+++ src/shared/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useProfileData'
 export * from './useInfiniteScroll'
